Исправил тип данных по геолокации, так как приходит не Int

+++ app/src/main/java/com/robivan/binlist/data/api/model/CountryDTO.kt
@@ -10,7 +10,7 @@ data class CountryDTO(
     @SerializedName("currency")
     val currency: String?,
     @SerializedName("latitude")
-    val latitude: Int?,
+    val latitude: Double?,
     @SerializedName("longitude")
-    val longitude: Int?
+    val longitude: Double?
 )
